backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
dateparser==1.2.0
distro==1.9.0
Django==5.1.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.14.0
freezegun==1.5.1
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
icalendar==6.1.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
langcodes==3.4.1
langid==1.1.6
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.0.2
openai==1.54.3
packaging==24.2
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
srsly==2.4.8
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
wrapt==1.16.0
yarl==1.17.2
